backend/requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.8.2
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
requests==2.20.0
